+++ src/frontend_canister/src/Pages/DashboardPage/Student/course/AllCourses/AllCourses.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const AllCourses = () => {
+    return (
+        <div>AllCourses</div>
+    )
+}
+
+export default AllCourses
