+++ src/frontend_canister/src/Pages/DashboardPage/Student/course/MyCourses/AllCourses.jsx
@@ -1,0 +1,87 @@
+import React from 'react';
+import Surr from '../../../../../../assets/surr-809 2.png'
+const AllCourses = () => {
+    return (
+        <div className="container px-4 py-5 mx-auto font-poppins rounded-xl">
+            <div className="px-8 py-6 bg-white rounded-lg shadow-md">
+                <div className="flex items-center justify-between mb-4">
+                    <h2 className="text-2xl font-semibold">All Courses</h2>
+                    <h5 className="text-[#925FE2]">see all</h5>
+                </div>
+                <ul>
+                    <li className="flex items-center justify-between py-4">
+                        <div className="flex-shrink-0">
+                            <div className='bg-[#CBD1FF] rounded-lg p-2'>
+                                <img className="w-12 h-12 rounded-full" src={Surr} alt="Course thumbnail" />
+                            </div>
+                        </div>
+                        <div className="flex-grow ml-4">
+                            <h3 className="text-base font-medium">Blockchain Course</h3>
+                            <p className="text-sm text-gray-600">By Name</p>
+                            <div className="flex items-center text-sm text-gray-500">
+                                <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                    <path d="M16.0986 19.7634C15.95 19.763 15.8037 19.7265 15.6724 19.657L11.0001 17.2004L6.32785 19.657C6.17657 19.7363 6.00615 19.7717 5.83581 19.7593C5.66548 19.7469 5.502 19.6871 5.3638 19.5868C5.22561 19.4864 5.1182 19.3494 5.05369 19.1913C4.98918 19.0332 4.97013 18.8602 4.99868 18.6918L5.8906 13.4888L2.11027 9.80377C1.98797 9.68437 1.9015 9.53315 1.86061 9.3672C1.81971 9.20125 1.82603 9.02717 1.87885 8.86462C1.93167 8.70206 2.02888 8.55752 2.15951 8.4473C2.29014 8.33707 2.44898 8.26557 2.6181 8.24085L7.84218 7.48185L10.1788 2.74819C10.2629 2.60504 10.383 2.48637 10.5272 2.40391C10.6713 2.32145 10.8345 2.27808 11.0006 2.27808C11.1666 2.27808 11.3298 2.32145 11.4739 2.40391C11.6181 2.48637 11.7382 2.60504 11.8224 2.74819L14.158 7.48185L19.3821 8.24085C19.5512 8.26557 19.7101 8.33707 19.8407 8.4473C19.9713 8.55752 20.0685 8.70206 20.1213 8.86462C20.1742 9.02717 20.1805 9.20125 20.1396 9.3672C20.0987 9.53315 20.0122 9.68437 19.8899 9.80377L16.1096 13.4888L17.0024 18.6918C17.025 18.8233 17.0185 18.9582 16.9836 19.0869C16.9486 19.2157 16.8859 19.3353 16.7999 19.4373C16.714 19.5394 16.6067 19.6214 16.4858 19.6777C16.3648 19.7341 16.232 19.7633 16.0986 19.7634Z" fill="#FFAE63" />
+                                </svg>
+                                <span className='ml-1'>4.8 Rating</span>
+                            </div>
+                        </div>
+                    </li>
+                    <li className="flex items-center justify-between py-4">
+                        <div className="flex-shrink-0">
+                            <div className='bg-[#CBD1FF] rounded-lg p-2'>
+                                <img className="w-12 h-12 rounded-full" src={Surr} alt="Course thumbnail" />
+                            </div>
+                        </div>
+                        <div className="flex-grow ml-4">
+                            <h3 className="text-base font-medium">Blockchain Course</h3>
+                            <p className="text-sm text-gray-600">By Name</p>
+                            <div className="flex items-center text-sm text-gray-500">
+                                <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                    <path d="M16.0986 19.7634C15.95 19.763 15.8037 19.7265 15.6724 19.657L11.0001 17.2004L6.32785 19.657C6.17657 19.7363 6.00615 19.7717 5.83581 19.7593C5.66548 19.7469 5.502 19.6871 5.3638 19.5868C5.22561 19.4864 5.1182 19.3494 5.05369 19.1913C4.98918 19.0332 4.97013 18.8602 4.99868 18.6918L5.8906 13.4888L2.11027 9.80377C1.98797 9.68437 1.9015 9.53315 1.86061 9.3672C1.81971 9.20125 1.82603 9.02717 1.87885 8.86462C1.93167 8.70206 2.02888 8.55752 2.15951 8.4473C2.29014 8.33707 2.44898 8.26557 2.6181 8.24085L7.84218 7.48185L10.1788 2.74819C10.2629 2.60504 10.383 2.48637 10.5272 2.40391C10.6713 2.32145 10.8345 2.27808 11.0006 2.27808C11.1666 2.27808 11.3298 2.32145 11.4739 2.40391C11.6181 2.48637 11.7382 2.60504 11.8224 2.74819L14.158 7.48185L19.3821 8.24085C19.5512 8.26557 19.7101 8.33707 19.8407 8.4473C19.9713 8.55752 20.0685 8.70206 20.1213 8.86462C20.1742 9.02717 20.1805 9.20125 20.1396 9.3672C20.0987 9.53315 20.0122 9.68437 19.8899 9.80377L16.1096 13.4888L17.0024 18.6918C17.025 18.8233 17.0185 18.9582 16.9836 19.0869C16.9486 19.2157 16.8859 19.3353 16.7999 19.4373C16.714 19.5394 16.6067 19.6214 16.4858 19.6777C16.3648 19.7341 16.232 19.7633 16.0986 19.7634Z" fill="#FFAE63" />
+                                </svg>
+                                <span className='ml-1'>4.8 Rating</span>
+                            </div>
+                        </div>
+                    </li>
+                    <li className="flex items-center justify-between py-4">
+                        <div className="flex-shrink-0">
+                            <div className='bg-[#CBD1FF] rounded-lg p-2'>
+                                <img className="w-12 h-12 rounded-full" src={Surr} alt="Course thumbnail" />
+                            </div>
+                        </div>
+                        <div className="flex-grow ml-4">
+                            <h3 className="text-base font-medium">Blockchain Course</h3>
+                            <p className="text-sm text-gray-600">By Name</p>
+                            <div className="flex items-center text-sm text-gray-500">
+                                <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                    <path d="M16.0986 19.7634C15.95 19.763 15.8037 19.7265 15.6724 19.657L11.0001 17.2004L6.32785 19.657C6.17657 19.7363 6.00615 19.7717 5.83581 19.7593C5.66548 19.7469 5.502 19.6871 5.3638 19.5868C5.22561 19.4864 5.1182 19.3494 5.05369 19.1913C4.98918 19.0332 4.97013 18.8602 4.99868 18.6918L5.8906 13.4888L2.11027 9.80377C1.98797 9.68437 1.9015 9.53315 1.86061 9.3672C1.81971 9.20125 1.82603 9.02717 1.87885 8.86462C1.93167 8.70206 2.02888 8.55752 2.15951 8.4473C2.29014 8.33707 2.44898 8.26557 2.6181 8.24085L7.84218 7.48185L10.1788 2.74819C10.2629 2.60504 10.383 2.48637 10.5272 2.40391C10.6713 2.32145 10.8345 2.27808 11.0006 2.27808C11.1666 2.27808 11.3298 2.32145 11.4739 2.40391C11.6181 2.48637 11.7382 2.60504 11.8224 2.74819L14.158 7.48185L19.3821 8.24085C19.5512 8.26557 19.7101 8.33707 19.8407 8.4473C19.9713 8.55752 20.0685 8.70206 20.1213 8.86462C20.1742 9.02717 20.1805 9.20125 20.1396 9.3672C20.0987 9.53315 20.0122 9.68437 19.8899 9.80377L16.1096 13.4888L17.0024 18.6918C17.025 18.8233 17.0185 18.9582 16.9836 19.0869C16.9486 19.2157 16.8859 19.3353 16.7999 19.4373C16.714 19.5394 16.6067 19.6214 16.4858 19.6777C16.3648 19.7341 16.232 19.7633 16.0986 19.7634Z" fill="#FFAE63" />
+                                </svg>
+                                <span className='ml-1'>4.8 Rating</span>
+                            </div>
+                        </div>
+                    </li>
+                    <li className="flex items-center justify-between py-4">
+                        <div className="flex-shrink-0">
+                            <div className='bg-[#CBD1FF] rounded-lg p-2'>
+                                <img className="w-12 h-12 rounded-full" src={Surr} alt="Course thumbnail" />
+                            </div>
+                        </div>
+                        <div className="flex-grow ml-4">
+                            <h3 className="text-base font-medium">Blockchain Course</h3>
+                            <p className="text-sm text-gray-600">By Name</p>
+                            <div className="flex items-center text-sm text-gray-500">
+                                <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                    <path d="M16.0986 19.7634C15.95 19.763 15.8037 19.7265 15.6724 19.657L11.0001 17.2004L6.32785 19.657C6.17657 19.7363 6.00615 19.7717 5.83581 19.7593C5.66548 19.7469 5.502 19.6871 5.3638 19.5868C5.22561 19.4864 5.1182 19.3494 5.05369 19.1913C4.98918 19.0332 4.97013 18.8602 4.99868 18.6918L5.8906 13.4888L2.11027 9.80377C1.98797 9.68437 1.9015 9.53315 1.86061 9.3672C1.81971 9.20125 1.82603 9.02717 1.87885 8.86462C1.93167 8.70206 2.02888 8.55752 2.15951 8.4473C2.29014 8.33707 2.44898 8.26557 2.6181 8.24085L7.84218 7.48185L10.1788 2.74819C10.2629 2.60504 10.383 2.48637 10.5272 2.40391C10.6713 2.32145 10.8345 2.27808 11.0006 2.27808C11.1666 2.27808 11.3298 2.32145 11.4739 2.40391C11.6181 2.48637 11.7382 2.60504 11.8224 2.74819L14.158 7.48185L19.3821 8.24085C19.5512 8.26557 19.7101 8.33707 19.8407 8.4473C19.9713 8.55752 20.0685 8.70206 20.1213 8.86462C20.1742 9.02717 20.1805 9.20125 20.1396 9.3672C20.0987 9.53315 20.0122 9.68437 19.8899 9.80377L16.1096 13.4888L17.0024 18.6918C17.025 18.8233 17.0185 18.9582 16.9836 19.0869C16.9486 19.2157 16.8859 19.3353 16.7999 19.4373C16.714 19.5394 16.6067 19.6214 16.4858 19.6777C16.3648 19.7341 16.232 19.7633 16.0986 19.7634Z" fill="#FFAE63" />
+                                </svg>
+                                <span className='ml-1'>4.8 Rating</span>
+                            </div>
+                        </div>
+                    </li>
+                    {/* Additional list items go here */}
+                </ul>
+            </div>
+        </div>
+    );
+}
+
+export default AllCourses;
