+++ src/frontend_canister/src/Pages/DashboardPage/Student/course/MyCourses/Features.jsx
@@ -1,0 +1,54 @@
+import React from 'react';
+
+const FeatureList = () => {
+    return (
+        <div className='container mx-auto px-4 py-8 font-poppins rounded-xl flex flex-wrap gap-4'>
+
+            <div className="bg-white shadow-md rounded-lg overflow-hidden flex flex-col justify-center items-center p-4 w-full md:w-1/2 lg:w-1/5">
+                <h2 className='text-gray-700 text-lg  mb-2'>Lessons</h2>
+                <div className="flex items-center">
+                    <svg className="w-6 h-6 mr-2" viewBox="0 0 20 23" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M17.4999 3.69531H2.49992C1.57492 3.69531 0.833252 4.43698 0.833252 5.36198V15.362C0.833252 15.804 1.00885 16.2279 1.32141 16.5405C1.63397 16.8531 2.05789 17.0286 2.49992 17.0286H6.66658V18.6953H13.3333V17.0286H17.4999C18.4166 17.0286 19.1583 16.2786 19.1583 15.362L19.1666 5.36198C19.1666 4.91995 18.991 4.49603 18.6784 4.18347C18.3659 3.87091 17.9419 3.69531 17.4999 3.69531ZM17.4999 15.362H2.49992V5.36198H17.4999V15.362ZM13.3333 10.362L7.49992 13.6953V7.02865L13.3333 10.362Z" fill="#2D2F31" />
+                    </svg>
+                    <p >128</p>
+                </div>
+            </div>
+
+            <div className="bg-white shadow-md rounded-lg overflow-hidden flex flex-col justify-center items-center p-4 w-full md:w-1/2 lg:w-1/5">
+                <h2 className='text-gray-700 text-lg  mb-2'>Duration</h2>
+                <div className="flex items-center">
+                    <svg className="w-6 h-6 mr-2" viewBox="0 0 21 21" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path fill-rule="evenodd" clip-rule="evenodd" d="M10.5035 0.873291C5.19753 0.873291 0.871582 5.19311 0.871582 10.4982C0.871582 15.8032 5.1984 20.1283 10.5035 20.1283C15.8085 20.1283 20.1266 15.8032 20.1266 10.4982C20.1266 5.19311 15.8085 0.873291 10.5035 0.873291ZM10.5035 2.62327C11.538 2.62154 12.5627 2.82406 13.5188 3.21922C14.4749 3.61438 15.3436 4.1944 16.0751 4.92602C16.8066 5.65764 17.3864 6.52647 17.7813 7.48266C18.1763 8.43886 18.3786 9.46362 18.3766 10.4982C18.3787 11.533 18.1765 12.5581 17.7817 13.5146C17.3868 14.4711 16.8071 15.3404 16.0757 16.0724C15.3443 16.8045 14.4756 17.385 13.5194 17.7807C12.5632 18.1764 11.5383 18.3795 10.5035 18.3783C9.46819 18.3796 8.44283 18.1766 7.48609 17.7811C6.52936 17.3855 5.66004 16.8052 4.92791 16.0732C4.19578 15.3412 3.61521 14.472 3.21946 13.5154C2.8237 12.5588 2.62052 11.5334 2.62156 10.4982C2.62156 6.13897 6.14426 2.62327 10.5035 2.62327ZM10.4895 4.361C10.3738 4.36213 10.2595 4.38618 10.1532 4.43176C10.0469 4.47734 9.95071 4.54354 9.87016 4.62654C9.7896 4.70953 9.7263 4.80767 9.68392 4.91529C9.64153 5.0229 9.6209 5.13785 9.62322 5.25349V10.4982C9.6237 10.6133 9.64691 10.7272 9.6915 10.8334C9.7361 10.9395 9.8012 11.0358 9.88309 11.1168L13.383 14.6185C13.464 14.7018 13.5607 14.7682 13.6676 14.8137C13.7745 14.8592 13.8894 14.883 14.0055 14.8837C14.1217 14.8843 14.2368 14.8619 14.3442 14.8176C14.4516 14.7733 14.5491 14.708 14.631 14.6257C14.7129 14.5433 14.7776 14.4454 14.8213 14.3378C14.865 14.2301 14.8868 14.1149 14.8855 13.9987C14.8842 13.8826 14.8597 13.7678 14.8136 13.6612C14.7675 13.5546 14.7006 13.4582 14.6168 13.3778L11.3749 10.1359V5.25349C11.3773 5.13623 11.3561 5.01969 11.3125 4.91081C11.2689 4.80192 11.2039 4.70291 11.1213 4.61965C11.0387 4.53639 10.9402 4.47059 10.8316 4.42616C10.7231 4.38173 10.6067 4.35957 10.4895 4.361Z" fill="black" />
+                    </svg>
+                    <p >5h 28m</p>
+                </div>
+            </div>
+
+            <div className="bg-white shadow-md rounded-lg overflow-hidden flex flex-col justify-center items-center p-4 w-full md:w-1/2 lg:w-1/5">
+                <h2 className='text-gray-700 text-lg  mb-2'>Level</h2>
+                <div className="flex items-center">
+                    <svg className="w-6 h-6 mr-2" viewBox="0 0 26 26" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M7.72578 12.9258H4.71721C4.16007 12.9258 3.71436 13.3715 3.71436 13.9287V20.0201C3.71436 20.5773 4.16007 21.023 4.71721 21.023H7.72578C8.28293 21.023 8.72864 20.5773 8.72864 20.0201V13.9287C8.72864 13.3715 8.28293 12.9258 7.72578 12.9258ZM14.5229 10.5858H11.5144C10.9572 10.5858 10.5115 11.0315 10.5115 11.5887V20.0201C10.5115 20.5773 10.9572 21.023 11.5144 21.023H14.5229C15.0801 21.023 15.5258 20.5773 15.5258 20.0201V11.5887C15.4886 11.0315 15.0429 10.5858 14.5229 10.5858ZM21.2829 5.0144H18.2744C17.7172 5.0144 17.2715 5.46012 17.2715 6.01726V20.0201C17.2715 20.5773 17.7172 21.023 18.2744 21.023H21.2829C21.8401 21.023 22.2858 20.5773 22.2858 20.0201V5.98012C22.2858 5.46012 21.8401 5.0144 21.2829 5.0144Z" fill="black" />
+                    </svg>
+                    <p >Beginner</p>
+                </div>
+            </div>
+
+            <div className="bg-white shadow-md rounded-lg overflow-hidden flex flex-col justify-center items-center p-4 w-full md:w-1/2 lg:w-1/5">
+                <h2 className='text-gray-700 text-lg  mb-2'>Views</h2>
+                <div className="flex items-center">
+                    <svg className="w-6 h-6 mr-2" viewBox="0 0 32 32" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M16 23.6801C9.17825 23.6801 5.76001 16.8551 5.76001 16.0001C5.76001 15.145 9.17825 8.32007 16 8.32007C22.8218 8.32007 26.24 15.145 26.24 16.0001C26.24 16.8551 22.8218 23.6801 16 23.6801ZM7.04929 16.0001C7.28801 16.7889 10.2394 22.4001 16 22.4001C21.7606 22.4001 24.712 16.7889 24.9507 16.0001C24.712 15.2113 21.7606 9.60007 16 9.60007C10.2394 9.60007 7.28801 15.2113 7.04929 16.0001ZM16 20.4801C13.5299 20.4801 11.52 18.4705 11.52 16.0001C11.52 13.5297 13.5299 11.5201 16 11.5201C18.4701 11.5201 20.48 13.5297 20.48 16.0001C20.48 18.4705 18.4701 20.4801 16 20.4801ZM16 12.8001C14.2355 12.8001 12.8 14.2356 12.8 16.0001C12.8 17.7645 14.2355 19.2001 16 19.2001C17.7645 19.2001 19.2 17.7645 19.2 16.0001C19.2 14.2356 17.7645 12.8001 16 12.8001Z" fill="black" />
+                    </svg>
+                    <p>12456</p>
+                </div>
+            </div>
+        </div>
+
+
+
+
+    );
+};
+
+export default FeatureList;
