+++ src/frontend_canister/src/Pages/DashboardPage/Student/course/MyCourses/LearningObjectives.jsx
@@ -1,0 +1,110 @@
+import React, { useState } from 'react';
+
+function LearningObjectives() {
+
+
+    return (
+        <div className="container mx-auto px-4 py-8 font-poppins rounded-xl">
+    <div className="bg-white rounded-lg shadow-md px-8 py-6">
+        <h2 className='font-bold text-lg mb-4'>What you will learn</h2>
+        <ul className="space-y-4">
+            <li className="flex items-center">
+                <svg className="w-6 h-7 mr-3" viewBox="0 0 23 28" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <g clip-path="url(#clip0_110_2789)">
+                        <path d="M8.62491 17.9727L4.62866 14L3.26782 15.3433L8.62491 20.6688L20.1249 9.23659L18.7737 7.89331L8.62491 17.9727Z" fill="#2D2F31" />
+                    </g>
+                    <defs>
+                        <clip-path id="clip0_110_2789">
+                            <rect width="23" height="28" fill="white" />
+                        </clip-path>
+                    </defs>
+                </svg>
+                <span>Understand the theory behind Blockchain</span>
+            </li>
+            <li className="flex items-center">
+                <svg className="w-6 h-7 mr-3" viewBox="0 0 23 28" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <g clip-path="url(#clip0_110_2789)">
+                        <path d="M8.62491 17.9727L4.62866 14L3.26782 15.3433L8.62491 20.6688L20.1249 9.23659L18.7737 7.89331L8.62491 17.9727Z" fill="#2D2F31" />
+                    </g>
+                    <defs>
+                        <clip-path id="clip0_110_2789">
+                            <rect width="23" height="28" fill="white" />
+                        </clip-path>
+                    </defs>
+                </svg>
+                <span>Understand the theory behind Cryptocurrency Transactions</span>
+            </li>
+            <li className="flex items-center">
+                <svg className="w-6 h-7 mr-3" viewBox="0 0 23 28" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <g clip-path="url(#clip0_110_2789)">
+                        <path d="M8.62491 17.9727L4.62866 14L3.26782 15.3433L8.62491 20.6688L20.1249 9.23659L18.7737 7.89331L8.62491 17.9727Z" fill="#2D2F31" />
+                    </g>
+                    <defs>
+                        <clip-path id="clip0_110_2789">
+                            <rect width="23" height="28" fill="white" />
+                        </clip-path>
+                    </defs>
+                </svg>
+                <span>Understand the theory behind Smart Contracts</span>
+            </li>
+            <li className="flex items-center">
+                <svg className="w-6 h-7 mr-3" viewBox="0 0 23 28" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <g clip-path="url(#clip0_110_2789)">
+                        <path d="M8.62491 17.9727L4.62866 14L3.26782 15.3433L8.62491 20.6688L20.1249 9.23659L18.7737 7.89331L8.62491 17.9727Z" fill="#2D2F31" />
+                    </g>
+                    <defs>
+                        <clip-path id="clip0_110_2789">
+                            <rect width="23" height="28" fill="white" />
+                        </clip-path>
+                    </defs>
+                </svg>
+                <span>Create their own currency</span>
+            </li>
+            <li className="flex items-center">
+                <svg className="w-6 h-7 mr-3" viewBox="0 0 23 28" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <g clip-path="url(#clip0_110_2789)">
+                        <path d="M8.62491 17.9727L4.62866 14L3.26782 15.3433L8.62491 20.6688L20.1249 9.23659L18.7737 7.89331L8.62491 17.9727Z" fill="#2D2F31" />
+                    </g>
+                    <defs>
+                        <clip-path id="clip0_110_2789">
+                            <rect width="23" height="28" fill="white" />
+                        </clip-path>
+                    </defs>
+                </svg>
+                <span>Create their own Smart Contracts</span>
+            </li>
+            <li className="flex items-center">
+                <svg className="w-6 h-7 mr-3" viewBox="0 0 23 28" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <g clip-path="url(#clip0_110_2789)">
+                        <path d="M8.62491 17.9727L4.62866 14L3.26782 15.3433L8.62491 20.6688L20.1249 9.23659L18.7737 7.89331L8.62491 17.9727Z" fill="#2D2F31" />
+                    </g>
+                    <defs>
+                        <clip-path id="clip0_110_2789">
+                            <rect width="23" height="28" fill="white" />
+                        </clip-path>
+                    </defs>
+                </svg>
+                <span>Create their own Smart Contracts</span>
+            </li>
+            <li className="flex items-center">
+                <svg className="w-6 h-7 mr-3" viewBox="0 0 23 28" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <g clip-path="url(#clip0_110_2789)">
+                        <path d="M8.62491 17.9727L4.62866 14L3.26782 15.3433L8.62491 20.6688L20.1249 9.23659L18.7737 7.89331L8.62491 17.9727Z" fill="#2D2F31" />
+                    </g>
+                    <defs>
+                        <clip-path id="clip0_110_2789">
+                            <rect width="23" height="28" fill="white" />
+                        </clip-path>
+                    </defs>
+                </svg>
+                <span>Create their own Smart Contracts</span>
+            </li>
+        </ul>
+    </div>
+</div>
+
+    
+    );
+}
+
+export default LearningObjectives;
