--- src/frontend_canister/src/Pages/DashboardPage/Student/courses/AllCourses/AllCourses.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const AllCourses = () => {
-    return (
-        <div>AllCourses</div>
-    )
-}
-
-export default AllCourses
